requirements.txt
crawl4ai>=0.3.50
python-dotenv>=1.0.0
pydantic>=2.0.0

